pref: Move SettingsView to a separate file
